requirements.txt
Django==1.10.5
djangorestframework==3.5.3
Pillow==3.4.2
psycopg2==2.6.2
wheel==0.24.0
